refactor(domain): 修复了 BaseRepository 接口命名不规范的问题
