+++ app/domain/__init__.py
@@ -1,0 +1,2 @@
+from .user import User, UserBaseRepository
+from .comment import Comment, CommentBaseRepository
